requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==5.0.1
attrs==24.3.0
biopython==1.85
Bottleneck==1.4.2
brotlicffi==1.0.9.2
causal-conv1d==1.5.0.post8
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
conda-pack==0.8.1
contourpy==1.3.1
cycler==0.11.0
datasets==3.3.2
dgl-cu113==0.9.1.post1
dill==0.3.8
einops==0.8.1
fair-esm==2.0.0
filelock==3.17.0
fonttools==4.60.1
frozenlist==1.5.0
fsspec==2024.12.0
gmpy2==2.2.1
hf-xet==1.1.8
huggingface_hub==0.34.4
idna==3.7
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.8
logzero==1.7.0
mamba-ssm==2.2.4
MarkupSafe==3.0.2
matplotlib==3.10.6
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
networkx==3.4.2
ninja==1.13.0
numexpr==2.11.0
numpy==1.26.4
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pip==25.2
propcache==0.3.1
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.0
pycares==4.10.0
pycparser==2.23
pyparsing==3.2.0
PyQt6==6.9.1
PyQt6_sip==13.10.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.5
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.12.0
setuptools==78.1.1
sip==6.12.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.5.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.51.3
triton==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvloop==0.21.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.18.0
